fix: cardList 페이지 초기 필터 상세 목록으로 이동
